+++ static/css/message.css
@@ -1,0 +1,5 @@
+.rounded-box {
+    padding: 0.5rem 1rem;
+    border-radius: 0.5rem;
+    border: 1px solid black;
+}
