+++ view/components/footer.php
@@ -3,11 +3,20 @@
 ?>
 
 <button type="button" id="arrow_to_top" class="arrow_to_top">↑</button>
-<script src="static/js/arrow_to_top.js"></script>
 
 <footer>
     <p>G4rden - Made with ❤️ by <a target="_blank" href="https://github.com/nansphilip/G4rden">G4rden's Team.</a></p>
 </footer>
+
+<?php // Scripts 
+?>
+<script src="static/js/global.js"></script>
+<script src="static/js/arrow_to_top.js"></script>
+<?php if (is_array(App::$jsFiles)) {
+    foreach (App::$jsFiles as $file) {
+        echo "<script src='$file' type='module'></script>";
+    }
+} ?>
 
 </body>
