+++ view/home.php
@@ -4,15 +4,16 @@
 
 <?php
 // Includes the header
-require_once ("view/components/header.php");
+require_once("view/components/header.php");
 ?>
 
 <main>
     <h2><?= App::$pageTitle; ?></h2>
     <p>Welcome to G4rden. <a href="index.php?p=message">Click here to chat.</a></p>
+    <p style="height: 2000px;">f</p>
 </main>
 
 <?php
 // Includes the footer
-require_once ("view/components/footer.php");
+require_once("view/components/footer.php");
 ?>
